requirements.txt
selenium>=4.15.0
python-docx>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
openai>=1.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
undetected-chromedriver>=3.5.4
